requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.5
django-celery==3.1.17
djangorestframework==3.3.3
facebook-sdk==1.0.0
kombu==3.0.35
oauthlib==1.0.3
Pillow==2.9.0
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.16
pytz==2016.3
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
SQLAlchemy==1.0.12
tweepy==3.5.0
wheel==0.24.0
